% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getbbox_from_zones}
\alias{getbbox_from_zones}
\title{Get bounding box from zones}
\usage{
getbbox_from_zones(zones_file = "input/zones.geojson")
}
\arguments{
\item{zones_file}{Path to the GeoJSON file containing zones. Default is 'input/zones.geojson'.}
}
\value{
A character string representing the bounding box in the format "xmin, ymin, xmax, ymax".
}
\description{
This function reads a GeoJSON file containing zones and calculates the bounding box of the zones.
}
