% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make_elevation}
\alias{make_elevation}
\title{Get elevation data}
\usage{
make_elevation(
  url = NULL,
  file = "UK-dem-50m-4326.tif.gz",
  base_url = "https://play.abstreet.org/dev/data/input/shared/elevation/"
)
}
\arguments{
\item{url}{Full URL of the elevation dataset if available}

\item{file}{File name if hosted on a known site}

\item{base_url}{Base URL associated with the 'file' argument}
}
\description{
This function downloads elevation data from a source such as
https://play.abstreet.org/dev/data/input/shared/elevation/UK-dem-50m-4326.tif.gz
or https://assets.od2net.org/input/LisboaIST_10m_4326.tif
}
