% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make_origins}
\alias{make_origins}
\title{make_origins function}
\usage{
make_origins(
  osm_file = "input/input.osm.pbf",
  query = "SELECT osm_id FROM multipolygons WHERE building IS NOT NULL",
  output_file = "input/buildings.geojson"
)
}
\arguments{
\item{osm_file}{The file path or name of the OSM file in PBF format. Default is "input/input.osm.pbf".}

\item{query}{The SQL query to select the multipolygons with a non-null building attribute. Default is "SELECT osm_id FROM multipolygons WHERE building IS NOT NULL".}

\item{output_file}{The file path or name of the output GeoJSON file containing the centroids of the selected buildings. Default is "input/buildings.geojson".}
}
\value{
None
}
\description{
This function reads an OpenStreetMap file, selects the multipolygons with a non-null building attribute,
calculates the centroids of the selected buildings, and writes the resulting centroids to a GeoJSON file.
}
