% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make_osm}
\alias{make_osm}
\title{make_osm Function}
\usage{
make_osm(
  force_download = FALSE,
  zones_file = "input/zones.geojson",
  output_file = "input/input.osm.pbf"
)
}
\arguments{
\item{force_download}{A logical value indicating whether to force the download of OSM data even if it already exists. Default is \code{FALSE}.}

\item{zones_file}{The file path or name of the zones file in GeoJSON format. Default is \code{"input/zones.geojson"}.}

\item{output_file}{The file path or name of the output OSM file in PBF format. Default is "input/input.osm.pbf".}
}
\value{
This function does not return any value. It downloads and extracts OSM data based on the specified zones.
}
\description{
This function is used to download and extract OpenStreetMap (OSM) data based on specified zones.
}
\examples{
if (file.exists("input/zones.geojson")) {
  make_osm(force_download = TRUE, zones_file = "input/zones.geojson")
}
}
