% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make_zones}
\alias{make_zones}
\title{Generate a 'zones.geojson' file}
\usage{
make_zones(file)
}
\arguments{
\item{file}{Location or URL of zones file}
}
\description{
This function requires a zones file, e.g.
"https://raw.githubusercontent.com/nptscot/npt/main/data-raw/zones_edinburgh.geojson"
or a file on your computer.
It will generate a file in the input/ folder
}
